requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
asyncpg==0.30.0
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
iso8601==2.1.0
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pypika-tortoise==0.3.2
python-dotenv==1.0.1
pytz==2024.2
tortoise-orm==0.22.2
typing_extensions==4.12.2
yarl==1.18.3
